backend/requirements.txt
asgiref==3.8.1
astroid==3.3.4
cfgv==3.4.0
coverage==7.6.1
dill==0.3.9
distlib==0.3.8
Django==5.1.1
djangorestframework==3.15.2
filelock==3.16.1
identify==2.6.1
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
nodeenv==1.9.1
packaging==24.1
platformdirs==4.3.6
pluggy==1.5.0
pre-commit==3.8.0
pylint==3.3.1
pytest==8.3.3
pytest-django==4.9.0
PyYAML==6.0.2
sqlparse==0.5.1
tomlkit==0.13.2
virtualenv==20.26.6
